# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SSH Commander API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upgrades:
    get:
      tags: [upgrades]
      summary: List upgrade runs (newest first)
      operationId: listUpgradeRuns
      responses:
        "200":
          description: Upgrade runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UpgradeRun"
    post:
      tags: [upgrades]
      summary: Start an upgrade run (admin)
      operationId: createUpgradeRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeRunInput"
      responses:
        "201":
          description: Run created and started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRun"
  /upgrades/{id}:
    get:
      tags: [upgrades]
      summary: Get an upgrade run with its per-device tasks
      operationId: getUpgradeRun
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Run with tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRunWithTasks"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [upgrades]
      summary: Delete a finished upgrade run (admin)
      operationId: deleteUpgradeRun
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /upgrades/{id}/cancel:
    post:
      tags: [upgrades]
      summary: Cancel pending tasks of a running upgrade (admin)
      operationId: cancelUpgradeRun
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRun"

  # ─── Alerts & Notifications (2.0.0) ──────────────────────────────────
...
  /tokens:
    get:
      tags: [tokens]
      summary: List own API tokens (admins may pass all=true for everyone's)
      operationId: listApiTokens
      parameters:
        - name: all
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: Tokens (hashes never returned)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiToken"
    post:
      tags: [tokens]
      summary: Create an API token (plaintext shown exactly once)
      operationId: createApiToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiTokenInput"
      responses:
        "201":
          description: Created token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTokenCreated"
  /tokens/{id}:
    delete:
      tags: [tokens]
      summary: Revoke an API token (owner or admin)
      operationId: revokeApiToken
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Job Output Analytics (2.0.0) ────────────────────────────────────
  /jobs/output-search:
    get:
      tags: [jobs]
      summary: Full-text search across historical job task outputs
      operationId: searchJobOutputs
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer }
        - name: offset
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Matching tasks with context snippets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutputSearchPage"
  /jobs/{id}/output-groups:
    get:
      tags: [jobs]
      summary: Group a job's device outputs by identical content
      operationId: getJobOutputGroups
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Output groups, largest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutputGroup"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    UpgradeRun:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        status: { type: string, enum: [running, completed, failed, cancelled] }
        preBackup: { type: boolean }
        totalTasks: { type: integer }
        completedTasks: { type: integer }
        failedTasks: { type: integer }
        createdBy: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
      required: [id, name, status, preBackup, totalTasks, completedTasks, failedTasks, createdAt]
    UpgradeTask:
      type: object
      properties:
        id: { type: integer }
        runId: { type: integer }
        routerId: { type: ["integer", "null"] }
        routerName: { type: string }
        routerIp: { type: string }
        status: { type: string, enum: [pending, backing_up, upgrading, rebooting, verifying, success, failed, skipped] }
        oldVersion: { type: ["string", "null"] }
        newVersion: { type: ["string", "null"] }
        log: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        startedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
      required: [id, runId, routerName, routerIp, status]
    UpgradeRunWithTasks:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        status: { type: string, enum: [running, completed, failed, cancelled] }
        preBackup: { type: boolean }
        totalTasks: { type: integer }
        completedTasks: { type: integer }
        failedTasks: { type: integer }
        createdBy: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/UpgradeTask"
      required: [id, name, status, preBackup, totalTasks, completedTasks, failedTasks, createdAt, tasks]
    UpgradeRunInput:
      type: object
      properties:
        name: { type: string }
        routerIds:
          type: array
          items: { type: integer }
        groupIds:
          type: array
          items: { type: integer }
        preBackup:
          type: boolean
          description: Snapshot each device's config before upgrading (default true)

    # ─── Alerts & Notifications (2.0.0) ─────────────────────────────────
...
    ApiToken:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        username: { type: ["string", "null"] }
        name: { type: string }
        prefix: { type: string }
        scope: { type: string, enum: [read, write] }
        lastUsedAt: { type: ["string", "null"], format: date-time }
        expiresAt: { type: ["string", "null"], format: date-time }
        revokedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, userId, name, prefix, scope, createdAt]
    ApiTokenInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        scope: { type: string, enum: [read, write] }
        expiresDays:
          type: integer
          description: Days until expiry; omit for a non-expiring token
      required: [name, scope]
    ApiTokenCreated:
      type: object
      properties:
        token:
          type: string
          description: Full plaintext token — displayed exactly once at creation
        apiToken:
          $ref: "#/components/schemas/ApiToken"
      required: [token, apiToken]

    # ─── Job Output Analytics (2.0.0) ───────────────────────────────────
    OutputGroup:
      type: object
      properties:
        outputHash: { type: string }
        count: { type: integer }
        routerNames:
          type: array
          items: { type: string }
        taskIds:
          type: array
          items: { type: integer }
        statuses:
          type: array
          items: { type: string }
        sampleOutput: { type: string }
      required: [outputHash, count, routerNames, taskIds, statuses, sampleOutput]
    OutputSearchHit:
      type: object
      properties:
        jobId: { type: integer }
        jobName: { type: string }
        taskId: { type: integer }
        routerName: { type: string }
        routerIp: { type: ["string", "null"] }
        status: { type: string }
        snippet: { type: string }
        createdAt: { type: ["string", "null"], format: date-time }
      required: [jobId, jobName, taskId, routerName, status, snippet]
    OutputSearchPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OutputSearchHit"
        total: { type: integer }
      required: [items, total]